* Expliziter Typecast in blech.cpp

+++ Alfa-Back-Console/Blech.cpp
@@ -8,8 +8,8 @@ int Blech::getBlechAnzahl(double plaetzchenXGroesse, double plaetzchenYGroesse)
 {
 	double plaetzchenXGroesseMitAbstand = plaetzchenXGroesse + 40;
 	double plaetzchenYGroesseMitAbstand = plaetzchenYGroesse + 40;
-	int spaltenAnzahl=xGroesse / plaetzchenXGroesseMitAbstand;
-	int reihenAnzahl=yGroesse/ plaetzchenYGroesseMitAbstand;
+	int spaltenAnzahl = (int)(xGroesse / plaetzchenXGroesseMitAbstand);
+	int reihenAnzahl = (int)(yGroesse/ plaetzchenYGroesseMitAbstand);
 	return  spaltenAnzahl * reihenAnzahl;
 	
 }
